Include unit tests for errors.rs resources inside the module
